Coefficients of daughter states are now tracked.  If daughter state coefficients add to zero, then the daughter state is removed (though I believe if there is another daughter state found afterward, it will be added just fine, i.e. if |2>=-|1>, coefficient of |1> is set to 0, and is removed from daughter array, but if |3> = |1> then the daughter array will be |1> ).
If the state is eligible for being a basis vector (i.e. none of its daughters are found to already be in the basis) but the size of the daughter 
array is 0, then the state will not be added to the basis array.

Now, I believe, there is just Hamiltonian Generation to worry about.